+++ CMakeLists.txt
@@ -208,7 +208,9 @@ endif()
 add_link_options(--sysroot ${CMAKE_BINARY_DIR}/Root)
 
 include_directories(Userland/Libraries/LibC)
+include_directories(Userland/Libraries/LibCrypt)
 include_directories(Userland/Libraries/LibM)
+include_directories(Userland/Libraries/LibPthread)
 include_directories(Userland/Libraries/LibSystem)
 include_directories(Userland/Services)
 include_directories(Userland)
