+++ Userland/Libraries/LibCore/Account.cpp
@@ -9,6 +9,7 @@
 #include <AK/ScopeGuard.h>
 #include <LibCore/Account.h>
 #include <LibCore/File.h>
+#include <crypt.h>
 #include <errno.h>
 #include <grp.h>
 #include <pwd.h>
