+++ Userland/Libraries/LibCrypt/crypt.cpp
@@ -6,8 +6,9 @@
 #include <AK/Base64.h>
 #include <AK/Types.h>
 #include <LibCrypto/Hash/SHA2.h>
+#include <crypt.h>
+#include <errno.h>
 #include <string.h>
-#include <unistd.h>
 
 extern "C" {
